Force connections to close after workers complete their work.
After the multiprocessing pool is done the connection that the parent
process had established is stale.  For whatever reason, django is unable
to detect this, and tries to reuse the connection, and gets an error on
the next database query.

I am forcing all DB connections to close immediately after the proc pool
is closed, which means that it establishes a new connection before the
next query.